requirements.txt
# local package
-e .

# external requirements
flake8
python-dotenv>=0.5.1
# cmake is necessary to install PyPortfolioOpt
Cython
cmake
PyPortfolioOpt
pandas-datareader
sklearn
yfinance
